requirements.txt
aiohttp==3.8.0
aiohttp_apispec==2.2.1
aiomysql==0.0.21
aioredis==2.0.0
environs==9.3.4
marshmallow==3.14.0
marshmallow_enum==1.5.1
markupsafe==2.0.1
numpy==1.21.3
opencv_python==4.5.4.58
Pillow==8.4.0
scipy==1.7.1
SQLAlchemy==1.3.24
